refactor: move transforms.py from credentials to common module
These are generic format utilities (kebab/camel case, dotted key
expansion) not specific to credentials. Moved to common/ so they
can be reused elsewhere.
