+++ application_sdk/credentials/__init__.py
@@ -69,7 +69,7 @@ from application_sdk.credentials.registry import (
 )
 from application_sdk.credentials.resolver import CredentialResolver
 from application_sdk.credentials.spec import AgentCredentialSpec
-from application_sdk.credentials.transforms import (
+from application_sdk.common.transforms import (
     camel_to_kebab,
     expand_dotted_keys,
     flatten_auth_section,
